add goals

+++ src/app/components/fuelings/fuelings.tsx
@@ -2,15 +2,19 @@ import React, { useEffect, useState } from 'react';
 import { Box, Button, Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle, Divider, Fab, IconButton, List, ListItem, ListItemText, Paper, TextField, Typography, useTheme } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
 import RemoveCircleIcon from '@material-ui/icons/RemoveCircle';
+import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
 import { Api } from '../../../api';
 import { useCommonStyles } from '../common-styles';
 import { useAlertMessage } from '../../providers/alert-provider';
+import { useConfirm } from 'material-ui-confirm';
 
 export const Fuelings: React.FC = () => {
     const commonClasses = useCommonStyles();
     const theme = useTheme();
     const alert = useAlertMessage();
+    const confirm = useConfirm();
 
+    const [open, setOpen] = React.useState(false);
     const [fuelings, setFuelings] = useState<Fueling[]>([]);
     const [newFueling, setNewFueling] = useState<Fueling>({
         fuelingId: 0,
@@ -25,43 +29,69 @@ export const Fuelings: React.FC = () => {
     }, []);
 
     const onClickAddFueling = () => {
-        setNewFueling(newFueling => ({
-            ...newFueling,
+        setNewFueling({
+            fuelingId: 0,
             name: '',
-        }));
+        });
         setOpen(true);
     }
-
-    const [open, setOpen] = React.useState(false);
 
     const handleClose = () => {
         setOpen(false);
     };
 
     const handleAddFueling = () => {
-        Api.Fueling.addFueling(newFueling)
-            .then(({ data }) => {
-                setFuelings(fuelings => {
-                    return [
-                        ...fuelings,
-                        data,
-                    ]
-                });
-            })
-            .catch(error => alert.addMessage(error))
-            .finally(() => setOpen(false));
+        if (newFueling.fuelingId === 0) {
+            Api.Fueling.addFueling(newFueling)
+                .then(({ data }) => {
+                    setFuelings(fuelings => {
+                        return [
+                            ...fuelings,
+                            data,
+                        ]
+                    });
+                })
+                .catch(error => alert.addMessage(error))
+                .finally(() => setOpen(false));
+        }
+        else {
+            Api.Fueling.updateFueling(newFueling.fuelingId, newFueling)
+                .then(() => {
+                    setFuelings(fuelings => {
+                        return [
+                            ...fuelings.filter(fueling => fueling.fuelingId !== newFueling.fuelingId),
+                            { ...newFueling },
+                        ]
+                    });
+                })
+                .catch(error => alert.addMessage(error))
+                .finally(() => setOpen(false));
+        }
+    }
+
+    const onClickEditFueling = (fuelingId: number) => {
+        const data = fuelings.find(fueling => fueling.fuelingId === fuelingId);
+        if (data) {
+            setNewFueling({ ...data })
+            setOpen(true);
+        }
     }
 
     const handleDeleteFueling = (fuelingId: number) => {
-        Api.Fueling.deleteFueling(fuelingId)
+        const fueling = fuelings.find(f => f.fuelingId === fuelingId);
+        confirm({ description: `Are you sure you want to delete ${fueling?.name}?`})
             .then(() => {
-                const idx = fuelings.findIndex(f => f.fuelingId === fuelingId);
-                setFuelings(fuelings => {
-                    return [...fuelings.slice(0, idx), ...fuelings.slice(idx + 1)];
-                });
+                Api.Fueling.deleteFueling(fuelingId)
+                    .then(() => {
+                        const idx = fuelings.findIndex(f => f.fuelingId === fuelingId);
+                        setFuelings(fuelings => {
+                            return [...fuelings.slice(0, idx), ...fuelings.slice(idx + 1)];
+                        });
+                    })
+                    .catch(error => alert.addMessage(error))
+                    .finally(() => setOpen(false));
             })
-            .catch(error => alert.addMessage(error))
-            .finally(() => setOpen(false));
+            .catch(() => null);
     }
 
     const onChangeNewFuelingName: React.ChangeEventHandler<HTMLInputElement> = event => {
@@ -94,10 +124,15 @@ export const Fuelings: React.FC = () => {
                                 <React.Fragment key={fueling.fuelingId}>
                                     <ListItem>
                                         <Box flexGrow={1} display="flex" alignItems="center">
-                                            <Box flexGrow={1}><ListItemText primary={fueling.name}/></Box>
-                                            <IconButton size="small" onClick={() => handleDeleteFueling(fueling.fuelingId)}>
-                                                <RemoveCircleIcon />
-                                            </IconButton>
+                                            <Box flexGrow={1}><ListItemText primary={fueling.name} /></Box>
+                                            <Box>
+                                                <IconButton size="small" onClick={() => onClickEditFueling(fueling.fuelingId)}>
+                                                    <EditOutlinedIcon />
+                                                </IconButton>
+                                                <IconButton size="small" color="secondary" onClick={() => handleDeleteFueling(fueling.fuelingId)}>
+                                                    <RemoveCircleIcon />
+                                                </IconButton>
+                                            </Box>
                                         </Box>
                                     </ListItem>
                                     <Divider />
@@ -107,11 +142,12 @@ export const Fuelings: React.FC = () => {
                     </List>
                 </Paper>
             </Box>
+
             <Dialog open={open} onClose={handleClose} aria-labelledby="form-dialog-title">
-                <DialogTitle id="form-dialog-title">Add fueling</DialogTitle>
+                <DialogTitle id="form-dialog-title">Fueling</DialogTitle>
                 <DialogContent>
                     <DialogContentText>
-                        To add a fueling, please enter your the name of the fueling here.
+                        Please enter your the name of the fueling here.
                     </DialogContentText>
                     <TextField
                         autoFocus
@@ -129,7 +165,7 @@ export const Fuelings: React.FC = () => {
                         Cancel
                     </Button>
                     <Button onClick={handleAddFueling} color="primary">
-                        Add
+                        Save
                     </Button>
                 </DialogActions>
             </Dialog>
